backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
pydantic-settings==2.6.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
email-validator==2.2.0
python-multipart==0.0.18
aiosqlite==0.20.0
stripe>=10.0.0
python-dotenv==1.0.1
sqlalchemy==2.0.35
psycopg2-binary==2.9.9
gunicorn==23.0.0
slowapi==0.1.9
structlog==25.1.0
redis==5.2.1
alembic==1.13.3
Jinja2==3.1.4

